Compare paths in platform independent way.
This commit allows platform independent config.rktd files.

Before this change, the submission directories needed to use the path
separator convention of the platform that the handin server should run
on. With this change, a config.rktd can use either forward or backward
slash and should work on all platforms.