Always show a warning icon
Most messages that checkers send, in particular all that current/future
checkers send, clearly call for a warning icon because they're
about (potential) bugs.

The first checker(s) also have positive messages, but "warning" might
still mean "please pay attention to this message".

We could also allow the checker to specify which kind of warning it
should use, but I believe that'd require a protocol change, so let's not
do that.