Pass [exists-ok? #t] to rename-file-or-directory
How could I miss this?

So apparently I missed this parameter *and* never sent two submissions
during testing. Probably because I knew the semantics of the `rename(2)`
syscall and its tricky details, and assumed `rename-file-or-directory`
was the same -- which again it was in complex details, but not in this
basic one.

So this bug was there in almost all variations of this fix, except the
ones using call-with-atomic-output-file` with a workaround.