Fix error in `grade.rktd`
Running the given correct `grade.rktd` file gives an error. You don't
strictly need to run these files, but it should still work.

The error is:
```
application: not a procedure;
 expected a procedure that can be applied to arguments
  given: #f
  arguments.:
```

That's because `[grading-finished #false]` expands to `(#false)`, which
tries to call the boolean `#false`. To fix this, produce instead a dummy
procedure.