lang-def/ -> utue-info1-lang-def/
This makes the directory name more unique. I'm not sure about
this (hence the separate commit), but using `lang-def` in a global
namespace for files that go on (some) tutors' computers seems a
questionable idea.

Possible TODO: rename/separate the other folders somehow, for similar
reasons. Again, that's for later.