Correct setup for `grade-template.rktd`
Imported files should be imported with an absolute path. This assumes
that one has installed the `handin` repo with `raco pkg install`, and
that it has installed a version with this file available (so a recent
enough version from the right branch, or a more recent version after
this PR is merged).